https://www.cactusprofiles.com/about-us/
Cactus Profiles Pvt. Ltd., is one of the leading manufacturer of Insulated Panels. We undertake turn-key projects from design, engineering, manufacturing and installation of Pre-fabricated Buildings, Homes and Cold Rooms.

Our company is part of the AKG group of companies based out of New Delhi, India. The company was incorporated in 2003.